// src/app/shared/domains/price-level.interface.ts
export interface PriceLevel {
  id?: string;
  name?: string;
  creationDate?: string;
  deletionDate?: string;
  dowingPartPF?: string;
  dowingPartPJ?: string;
  lastChangeDate?: string;
  professionalEvaluationHasPriceLevels?: string;
  professionalPartPF?: string;
  professionalPartPJ?: string;
  sessionValuePF?: string;
  sessionValuePJ?: string;
}